Move Chat to its own namespace
Also:
- Consistency on event methods naming
